Wire agent_repo_tools tests into CI
The per-team test matrix only runs software_engineering_team/tests, so the
new agent_repo_tools tests would not execute or be coverage-gated in CI.

- Add an agent_repo_tools paths filter and a batched matrix entry running
  agent_repo_tools/tests/ with --cov=agent_repo_tools at the 90% gate; add
  the F_AGENT_REPO_TOOLS env wiring and the unclassified-backend exclusion.
- Include agent_repo_tools in the software_engineering filter since the
  Senior SWE imports the shared inspection tools.
- Move the Strands-registration / schema-verbatim tests into the
  agent_repo_tools suite so they run in the new job; the run_implement
  wiring/prompt tests stay in coding_team.
